🔧 Debugging Signup (Register) Feature 🪲
BUG 1 (index.js): Fixed the MongoDB connection string. Initially, I mistakenly included the database name in the connection string, but only the MongoDB URL was required. 😎

BUG 2 (apiError.js): Corrected the parameter order in the ApiError class. The error message was initially placed after the status code as an array. Now, I’ve updated it so that the message follows the status code directly, making it more intuitive when calling the class. 😎

BUG 3 (ApiRes.js): Declared the success variable within the class scope, not just in the constructor. 😎

✨ New Feature: After uploading an image to Cloudinary, the image is now automatically removed from the server to save space. 🚀